extract the mock context
