+++ functions/prepare-check/index.spec.js
@@ -41,19 +41,7 @@ describe('prepare-check function', () => {
       }
     }
 
-    const logFunc = function (level = 'INFO') { /* console.log(`[${level.toUpperCase()}]: `, ...R.tail(Array.from(arguments))) */ }
-
-    const context = {
-      bindings: {},
-      log: (function () {
-        const logHandler = R.partial(logFunc, ['info'])
-        logHandler.verbose = R.partial(logFunc, ['verbose'])
-        logHandler.info = R.partial(logFunc, ['info'])
-        logHandler.error = R.partial(logFunc, ['error'])
-        return logHandler
-      }()), // whoo-hoo: a function with properties.  Your enthusiasm may vary.
-      done: function () {}
-    }
+    const context = require('../mock-context')
 
     it('validates a valid message', () => {
       spyOn(context, 'done')
